refactor: Restructure folders and files
