--- src/App.js
@@ -1,34 +1,0 @@
-import React, { useContext } from 'react';
-import logo from './logo.svg';
-import './App.css';
-import { AuthContext } from './context/AuthContext';
-
-function App() {
-  const { isLoggedIn, login, logout } = useContext(AuthContext);
-
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-      {isLoggedIn ? (
-        <button onClick={logout}>Logout</button>
-      ) : (
-        <button onClick={login}>Login</button>
-      )}
-    </div>
-  );
-}
-
-export default App;
